backend/pyproject.toml
[project]
name = "invoice-nudge"
version = "1.0.0"
description = "InvoiceNudge"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "bcrypt>=4.2.0",
    "pyjwt>=2.9.0",
    "resend>=2.5.0",
    "stripe>=8.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx", "ruff"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
